requirements.txt
# LangChain Core
langchain==0.1.20
langchain-core==0.1.49
langchain-community==0.0.34

# Integrações específicas
langchain-chroma==0.1.1
langchain-openai==0.1.2
langchain-mistralai==0.1.1

# Vector Store
chromadb==0.4.24

# Framework Web
streamlit==1.28.0

# Processamento de PDF
pypdf==3.17.0

# Environment variables
python-dotenv==1.0.0

# APIs
openai==1.3.0
mistralai==0.1.0

# Análise de dados (para recomendações futuras)
pandas==2.0.0
scikit-learn==1.3.0
numpy==1.24.0